StockCritique/CrewAI/requirements.txt
# requirements.txt
fastapi
uvicorn
redis
pydantic
pydantic_core
praw
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
huggingface-hub
openai
yfinance
pandas
requests
pytest
crewai
sentence-transformers
tokenizers
transformers
tqdm
torch
numpy
beautifulsoup4
python-multipart
python-dotenv
pypdf
